refactor(server): isolate connector inventory synchronization
