+++ services/server/src/app.ts
@@ -1,4 +1,4 @@
-import { ECDH, randomUUID } from "node:crypto";
+import { randomUUID } from "node:crypto";
 import { existsSync } from "node:fs";
 import { resolve } from "node:path";
 import { isDeepStrictEqual } from "node:util";
@@ -40,12 +40,16 @@ import { ConnectorOperationError, RelayHub, RelayUnavailableError } from "./rela
 import type { RelayBroker } from "./relay-broker.js";
 import {
   canonicalUserCode,
+  isP256PublicKey,
   pkceChallenge,
   randomToken,
   randomUserCode,
   safeEqual,
   tokenHash
 } from "./security.js";
+import {
+  collectionContractDescriptorSchema
+} from "./protocol-schemas.js";
 import {
   asSyncMutation,
   contractRequirements,
@@ -113,6 +117,9 @@ import { registerMirrorPairingRoutes } from "./features/mirrors/pairing-routes.j
 import {
   registerConnectorManagementRoutes
 } from "./features/connectors/management-routes.js";
+import {
+  registerConnectorInventoryRoutes
+} from "./features/connectors/inventory-routes.js";
 import { sessionToken } from "./platform/session-cookies.js";
 import { audit } from "./platform/audit-events.js";
 import { authorityUrl } from "./platform/authority-url.js";
@@ -129,30 +136,6 @@ const operationSchema = z.enum(COLLECTION_OPERATIONS);
 const DEVICE_GRANT_TYPE = "urn:ietf:params:oauth:grant-type:device_code";
 const DEVICE_AUTHORIZATION_SECONDS = 600;
 const DEVICE_POLL_INTERVAL_SECONDS = 5;
-const contractRequirementSchema = z.object({
-  id: z.string().trim().min(1).max(100),
-  version: z.string().regex(
-    /^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(?:-[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?(?:\+[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?$/
-  )
-}).strict();
-const digestSchema = z.string().regex(/^sha256:[0-9a-f]{64}$/);
-const collectionContractImplementationSchema = z.object({
-  type_name: z.string().min(1).max(100),
-  type_version: z.number().int().positive(),
-  type_path: z.string().min(1).max(500).optional(),
-  digest: digestSchema,
-  fields: z.record(z.string(), z.string().min(1)),
-  binding: z.record(z.string(), z.unknown()).optional()
-}).strict();
-const collectionContractDescriptorSchema = z.object({
-  contract_type: z.literal("record"),
-  id: contractRequirementSchema.shape.id,
-  version: contractRequirementSchema.shape.version,
-  digest: digestSchema,
-  schema: z.record(z.string(), z.unknown()),
-  binding_schema: z.record(z.string(), z.unknown()).optional(),
-  implementations: z.array(collectionContractImplementationSchema).min(1).max(100)
-}).strict();
 const encryptedRelayRequestSchema = z.object({
   type: z.literal("encrypted_operation_request"),
   protocol_version: z.literal(ENCRYPTED_RELAY_PROTOCOL_VERSION),
@@ -446,6 +429,7 @@ export async function buildApp(options: BuildOptions) {
     db: options.db,
     tailscaleAuth: options.tailscaleAuth
   });
+  registerConnectorInventoryRoutes(app, { db: options.db });
 
   app.post("/v1/authority-adoptions", async (request, reply) => {
     if (!options.hostedCollections || !options.hostedProvider) {
@@ -1604,194 +1588,6 @@ export async function buildApp(options: BuildOptions) {
         };
       }))
     };
-  });
-
-  app.post("/v1/connectors/sync", async (request, reply) => {
-    const connector = await requireConnector(request, reply, options.db);
-    if (!connector) return;
-    const input = z.object({
-      relay_public_key: z.string().min(80).max(200).refine(isP256PublicKey).optional(),
-      inventory_revision: z.number().int().positive().max(Number.MAX_SAFE_INTEGER),
-      collections: z.array(z.object({
-        id: z.uuid(),
-        display_name: z.string().min(1).max(200),
-        spec_version: z.string().min(1).max(30),
-        enabled: z.boolean(),
-        contracts: z.array(collectionContractDescriptorSchema).max(100).default([])
-      })).max(1_000)
-    }).parse(request.body);
-    if (new Set(input.collections.map((collection) => collection.id)).size !== input.collections.length) {
-      throw new RequestValidationError("A collection may appear only once in an inventory.");
-    }
-    const connection = await options.db.connect();
-    try {
-      await connection.query("BEGIN");
-      await connection.query("SELECT id FROM users WHERE id = $1 FOR UPDATE", [
-        connector.user_id
-      ]);
-      const accepted = await connection.query(
-        `UPDATE connectors SET
-           inventory_revision = $2,
-           relay_public_key = COALESCE($3, relay_public_key),
-           last_seen_at = now()
-         WHERE id = $1 AND inventory_revision < $2
-         RETURNING id`,
-        [connector.id, input.inventory_revision, input.relay_public_key ?? null]
-      );
-      if (!accepted.rows[0]) {
-        const current = await connection.query<{ inventory_revision: string | number }>(
-          "SELECT inventory_revision FROM connectors WHERE id = $1",
-          [connector.id]
-        );
-        await connection.query("COMMIT");
-        return {
-          accepted: false,
-          inventory_revision: Number(current.rows[0]?.inventory_revision ?? 0),
-          collections: []
-        };
-      }
-
-      const synchronized = [];
-      for (const collection of input.collections) {
-        const existing = await connection.query<{
-          id: string;
-          authority_state: "active" | "candidate" | "retired";
-          authority_epoch: string | number;
-        }>(
-          `SELECT id, authority_state, authority_epoch
-           FROM collections WHERE connector_id = $1 AND local_id = $2`,
-          [connector.id, collection.id]
-        );
-        const activeAuthority = await connection.query<{
-          id: string;
-          authority_epoch: string | number;
-        }>(
-          `SELECT id, authority_epoch FROM collections
-           WHERE user_id = $1 AND local_id = $2 AND authority_state = 'active'`,
-          [connector.user_id, collection.id]
-        );
-        const hosted = await connection.query<{
-          authority_state: "active" | "transferring" | "transferred";
-          authority_epoch: string | number;
-          transferred_collection_id: string | null;
-        }>(
-          `SELECT authority_state, authority_epoch, transferred_collection_id
-           FROM hosted_collections WHERE id = $1`,
-          [collection.id]
-        );
-        const hostedAccess = await resolveHostedCollectionAccess(
-          connection,
-          connector.user_id,
-          collection.id
-        );
-        const existingCollection = existing.rows[0];
-        const currentAuthority = activeAuthority.rows[0];
-        const hostedCollection = hostedAccess ? hosted.rows[0] : undefined;
-        const isActivatedTransfer = Boolean(
-          hostedCollection?.authority_state === "transferred"
-          && hostedCollection.transferred_collection_id
-          && hostedCollection.transferred_collection_id === existingCollection?.id
-        );
-        const authorityState: "active" | "candidate" = hostedCollection
-          ? (isActivatedTransfer ? "active" : "candidate")
-          : (currentAuthority && currentAuthority.id !== existingCollection?.id
-            ? "candidate"
-            : "active");
-        const authorityEpoch = Number(
-          hostedCollection?.authority_epoch
-          ?? currentAuthority?.authority_epoch
-          ?? existingCollection?.authority_epoch
-          ?? 1
-        );
-        const enabled = authorityState === "active" && collection.enabled;
-        const row = await connection.query<{
-          id: string;
-          local_id: string;
-          authority_state: "active" | "candidate" | "retired";
-          authority_epoch: string | number;
-        }>(
-          `INSERT INTO collections
-             (id, user_id, connector_id, local_id, display_name, spec_version, enabled,
-              reported_enabled, present, authority_state, authority_epoch, contracts,
-              last_inventory_revision)
-           VALUES ($1, $2, $3, $4, $5, $6, $7, $8, true, $9, $10, $11::jsonb, $12)
-           ON CONFLICT(connector_id, local_id) DO UPDATE SET
-             user_id = excluded.user_id,
-             display_name = excluded.display_name,
-             spec_version = excluded.spec_version,
-             enabled = excluded.enabled,
-             reported_enabled = excluded.reported_enabled,
-             present = true,
-             authority_state = excluded.authority_state,
-             authority_epoch = excluded.authority_epoch,
-             contracts = excluded.contracts,
-             last_inventory_revision = excluded.last_inventory_revision,
-             last_seen_at = now(),
-             removed_at = NULL
-           RETURNING id, local_id, authority_state, authority_epoch`,
-          [
-            randomUUID(),
-            connector.user_id,
-            connector.id,
-            collection.id,
-            collection.display_name,
-            collection.spec_version,
-            enabled,
-            collection.enabled,
-            authorityState,
-            authorityEpoch,
-            JSON.stringify(collection.contracts),
-            input.inventory_revision
-          ]
-        );
-        synchronized.push({
-          id: row.rows[0].local_id,
-          authority_state: row.rows[0].authority_state,
-          authority_epoch: Number(row.rows[0].authority_epoch)
-        });
-      }
-
-      const removed = await connection.query<{ id: string }>(
-        `UPDATE collections SET
-           present = false,
-           enabled = false,
-           authority_state = 'retired',
-           removed_at = now()
-         WHERE connector_id = $1 AND present = true
-           AND last_inventory_revision < $2
-         RETURNING id`,
-        [connector.id, input.inventory_revision]
-      );
-      for (const collection of removed.rows) {
-        const revoked = await connection.query<{ id: string }>(
-          `UPDATE grants SET revoked_at = COALESCE(revoked_at, now())
-           WHERE collection_id = $1 AND revoked_at IS NULL
-           RETURNING id`,
-          [collection.id]
-        );
-        for (const grant of revoked.rows) {
-          await connection.query(
-            "UPDATE access_tokens SET revoked_at = COALESCE(revoked_at, now()) WHERE grant_id = $1",
-            [grant.id]
-          );
-          await connection.query(
-            "UPDATE refresh_tokens SET revoked_at = COALESCE(revoked_at, now()) WHERE grant_id = $1",
-            [grant.id]
-          );
-        }
-      }
-      await connection.query("COMMIT");
-      return {
-        accepted: true,
-        inventory_revision: input.inventory_revision,
-        collections: synchronized
-      };
-    } catch (error) {
-      await connection.query("ROLLBACK");
-      throw error;
-    } finally {
-      connection.release();
-    }
   });
 
   app.post("/v1/connectors/collections/:collectionId/authority-transfers", async (request, reply) => {
@@ -6143,18 +5939,6 @@ async function rotateGrantEncryption(db: DatabasePool, grantId: string): Promise
     grantId,
     JSON.stringify(rotated)
   ]);
-}
-
-function isP256PublicKey(value: string): boolean {
-  if (!/^[A-Za-z0-9_-]+$/.test(value)) return false;
-  try {
-    const bytes = Buffer.from(value, "base64url");
-    if (bytes.length !== 65 || bytes[0] !== 4) return false;
-    ECDH.convertKey(bytes, "prime256v1", undefined, undefined, "uncompressed");
-    return true;
-  } catch {
-    return false;
-  }
 }
 
 function contractsSatisfy(
